Relax returned lifetime in PdfDocument::get_page
glib uses reference counting, and PopplerPage already owns a reference
to its PopplerDocument. So dropping our PdfDocumetn wrapper
after a PdfPage is created is fine.
However, the internal PopplerDocument still has a pointer
to the original bytes data, and that data needs to be kept
alive.

Return `PdfPage<'data>` instead of `PdfPage<'doc>` to reflect all this.